SERVER-91391 Split authLocalGetUser FP (#23317)
# Issue

SERVER-90391

# Description

`jstests/auth/user-cache-invalidation.js` depended on a FailPoint which
paused UserDocument resolution *after* the loading of the primary User
document, but before resolution of subroles, but this test expected that
the delay would occur prior to the user document load.

This change introduces a new failpoint which causes a delay prior to the
user document load.
Because of the previously misleading name of the existing failpoint, it
introduces the new one under the old name, and renames the old one to a
more descriptive identifier.