SERVER-89806 Remove kOutstandingAdditionalParticipant readOnly value (#23172)
# Issue

SERVER-89806

# Description
This patch:
1. Removes the "kOutstandingAdditionalParticipant" readOnly value from
the TransactionRouter. Instead, we now pass a flag into
TransactionRouter::processParticipantResponse to indicate if a response
is coming from a getMore request. If it is and the response contains an
additionalParticipant that the parent router does not know about, the
parent router will mark this shard as readOnly. This will allow the
shard to participate in the commit protocol as a read shard, even if the
client never receives data from this shard. Note that this also allows
this added shard to release its transaction resources during the commit,
rather than waiting for transaction lifetime limit secs.
2. Allows a sub-router to retry on
CommandOnShardedViewNotSupportedOnMongod. If a sub-router receives this
error, it will now clear it's pending participants (which will only
consist of the single node who threw the error), and then allow retrying
with the extra info provided in the error. It will NOT abort the
transaction on the shard who threw the error, but if this was the first
statement that the shard that threw the error ran, it will abort the
transaction itself. This means that if this wasn't the first statement,
the transaction may be leaked and hold onto transaction resources.
3. Allows a sub-router to change its clusterTime iff it receives a retry
of a request that "restarts" the transaction. Prior to the changes
described in #2, there was not a case where a parent router could
"restart" a transaction and pick a new clusterTime when a participant
acted as a sub-router, because we only allow restarting/picking a new
clusterTime on certain retryable errors, and we only allow retrying on
those if there was only one participant. If a shard had acted as a
sub-router, it would previously always include the shard it targeted
(regardless of error) in its response to the parent router, so the
parent router would never choose to retry/restart. Now that we clear the
participant on the sub-router after a
CommandOnShardedViewNotSupportedOnMongod error, it's possible that when
the shard retries, it gets an error like
ShardCannotRefreshDueToLocksHeld and responds to the parent router with
this error leading the parent router to retry and pick a new
clusterTime.
4. Disallows a shard from retrying on StaleConfig errors if the shard
has acted as a sub-router and has participants in its list.

# Testing
[Run
1](https://spruce.mongodb.com/version/6662fe0f98b3f30007327672/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)
[Run
2](https://spruce.mongodb.com/version/6662fe1ab71ab70007f42a5d/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)

The failures I'm seeing in the patches are 
1. BF-33160
2. A failure to find all the docs when running a $lookup in
agg_lookup.js, but in suites where we aren't running with transactions
(and this test doesn't run transactions) - the two suites I see this in
are
concurrency_sharded_with_stepdowns_and_balancer_and_config_transitions
and concurrency_sharded_initial_sync. I don't see existing BFs for this
issue, but since the test isn't running with transactions in these
failures it shouldn't be caused by this patch - I'm going to spend a bit
more time trying to understand what's up though.