SERVER-91245 Don't Check feature flag for data-bearing node stats repeatedly (#23115)
# Issue

SERVER-91245

# Description

[SERVER-91218](https://jira.mongodb.org/browse/SERVER-91218) requested
that we audit usage of this flag for dangerous patterns, including
accessing it more than once from the same thread. We do in fact do that
in a few cases related to aggregate commands. The solution is to push up
the decision of whether or not to request metrics as high as possible,
similar to how we do for find commands.

For cluster aggregate commands, this means we make the decision in
`ClusterAggregate::runAggregate` and pass a boolean down as far as
needed. Other entry points to aggregation functionality that also have
to make the decision are `targetShardsAndAddMergeCursors` and
`preparePipelineForExecution`.

I think the impact of this problem is fairly low - i.e., we might
request metrics for some parts of a query and not others, and therefore
undercount certain metrics for any queries that are sampled for query
stats during an unstable FCV period. It shouldn't cause a mongos to
request metrics from a mongod that doesn't support them, which would be
more serious.

# Testing

- [x] https://spruce.mongodb.com/version/6661cdd662a1720007c04187 -
failures are known issues