SERVER-91362 Perf: Don't copy JS scope obj if cached JsExecution exists (#23323)
# Issue

SERVER-91362 Perf: Don't copy JS scope obj if cached JsExecution exists

# Description

Starting in 5.0.0, an aggregation pipeline that needs a JsExecution
object always retrieves the JavaScript "scope" object every time this
object is used in order to pass it to JsExecution::get() in case the
JsExecution object has not already been created, as it is needed to
create a JsExecution object. However, in sharded clusters this can
result in making a new copy of the scope object for every document the
query processes, and if the scope object is large this can add a
significant overhead.

However, once the JsExecution object has been created for a given query
thread, it gets cached for the lifetime of that thread, so the scope
object is only needed once per thread. Thus there is no need to retrieve
and potentially make another copy of the "scope" object if there is
already a cached JsExecution object.

The problem was introduced in 5.0.0 by the following commit from
SERVER-47398:

https://github.com/mongodb/mongo/commit/a71a2a8bfecf7de0807a28e3eabf9412dddd4258

# Testing

I manually verified the fix on v5.0.23 where it reduced the original
107% performance regression vs 4.4.28 down to only a 3% regression,
using a 2-shard cluster with PSA replication and a query from the
[HELP-58993](https://jira.mongodb.org/browse/HELP-58993) customer who
surfaced this issue, using 100,000 synthetic input documents.

[Patch build
605](https://spruce.mongodb.com/version/66688ca13bf0cb00086a6bbe/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)
passed on this PR. The failures are unrelated:
- 4 known issues
- 1 unrelated config fuzzer failure

[Patch build
608](https://spruce.mongodb.com/version/6669ccbbbb57a70007ec4640/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)
passed on [commit
2](https://github.com/10gen/mongo/pull/23323/commits/e80cead705106f7157be4cfbe6c011a9d5e5d535)
of this PR. The failure are unrelated:
- 3 known issues
- 1 unrelated timeout while dropping a sharded database

<!--
Available Evergreen Comment Commands
evergreen patch - attempts to create a new PR patch that reruns on each
new commit; this is required to create a PR patch when only manual PR
testing is enabled
evergreen keep-definitions - reuse the tasks from the previous patch in
subsequent patches. Use this after you modify a patch's tasks.
evergreen reset-definitions - reset the patch tasks to the original
definition
evergreen refresh - resyncs PR GitHub checks
evergreen help - comments the above info
 -->