SERVER-91361 Approximate when it's safe to check for the updated value in non_transaction_snapshot_reads_retry.js. (#23271)
# Issue

SERVER-91361

# Description

There's no good way to pass `updateTS` to `read()`; even `TestData`
doesn't work because it's serialized before being passed to
`startParallelShell`, so values in `TestData` updated later aren't
propagated. There's no way for the shells to communicate even though we
want `read()` to know `updateTS` since it's not guaranteed safe to
`find()` until after the cluster time has advanced at least up to
`updateTS`. Therefore, we look for whether we are in the case of `>=
insertTS, < updateTS` or if we are in the case of `>= updateTS` and
check the response appropriately.

# Testing

- [x]
[Patch](https://spruce.mongodb.com/version/666a11344c136400079a5ae8/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)
~[Patch](https://spruce.mongodb.com/version/6669bf035b46be0007284dbb/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)~
~[Patch](https://spruce.mongodb.com/version/666752383d732400075069aa/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)~