SERVER-91352 Do not check for the no-op entry in readConcern_atClusterTime_noop_write.js (#23309)
# Issue
The test is trying to verify the transaction correctly advanced it's
lastOpTime before running a transaction with snapshot readConcern. This
is a corner case and can cause the transaction to hang indefinitely (or
until another operation advance its lastOpTime). The test tried to
verify a specific "no-op" was performed in the oplog, ensuring the
forced advanced of the lastOpTime.
However, such an approach is very fragile as any background operation
might perform write in the op-log. This will cause the test to fail
without having any specific issue

# Description
The test has been modified to verify the transaction correctly worked
without hanging.
The test has been refactored for readability and future testability.

# Testing


https://spruce.mongodb.com/version/666aabd74d3d1b0007524994/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC