SERVER-91318 Prevent config shard CatalogCache refresh deadlock (#23297)
# Issue
SERVER-91318

# Description
Prevent CatalogCache refresh deadlock on config shard nodes. The
deadlock can happen due to a node sending an aggregation to itself
during the refresh, which in turn requires a refresh. If the
CatalogCache thread pool is full, the node ends up in deadlock.

# Testing
Added a reproducer, which hangs before the change.

https://spruce.mongodb.com/version/66670c1f5dd0af0007b8a5a2/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC