SERVER-91483 Lower number of iterations in query_stats_concurrent.js (#23456)
The number of iterations for `multipleShapes()` can be changed to 2000
to evict an entry.

Each entry is conservatively estimated to be ~1k bytes and with a 1 MB
cache, we can estimate the number of entries needed to evict an entry to
be ~1050 queries. To be safe, I've doubled this value for the number of
iterations, which has successfully evicted in testing, and should avoid
the timeouts.