SERVER-91186: Switch to using RHEL 9.4 for performance testing. (#23425)
RHEL 7.0 is very end of life and Mongo is dropping support for it, so we
should move to the newest version.

This reverts commit fa860a121ac22369f987a7683bd098185e5133a5, which was
an erroneous reversion.

Anything in this description will be included in the commit message.
Replace or delete this text before merging. Add links to testing in the
comments of the PR.