SERVER-90733 Add parser defaults (#23516)
# Issue

Fixes bug in clang_tidy.py script of a missing variable.

# Description

Under certain scenarios, like if `compile_commands.json` doesn't exist,
the clang_tidy.py script fails to fallback to defaults and spits out a
stacktrace to the user. This was introduced by SERVER-90733 when it made
this script more modular. This was found when backporting and testing on
8.0 [here](https://github.com/10gen/mongo/pull/23514).