SERVER-91559 Fix bazel_run.sh's exit code (#23537)
Previously the exit code on `bazel_run.sh` and `bazel_build.sh` was
ignored due to a bug introduced a couple weeks ago in
https://jira.mongodb.org/browse/SERVER-90875

This only affects the evergreen tasks that were calling Bazel through
the shell script, and not the Bazel build ran to compile mongo code (run
through scons.py).

The two things this hid were linter failures and a clang-tidy error
introduced by enabling warnings-as-errors. The linter failures are
addressed here by running the formatter and the clang-tidy issue is
resolved in this PR by switching the compiler passed to clang-tidy from
gcc to clang.

This would affect the following checks:
- codeowners linting
- prettier linting
- clang-tidy on targets converted to bazel
- a few one-off unit tests setup for when we switch to invoking the full
build with "bazel build" instead of scons.py

Required PB:
https://spruce.mongodb.com/version/666ce36fa0d3e30007c214c1/tasks
Linter/Codeowners:
https://spruce.mongodb.com/version/666cdca5f4de0a000778f1ab/tasks
Linter/Codeowners (with formatting error added):
https://spruce.mongodb.com/version/666cdd6698ba280007cb66be/tasks