SERVER-91262: Move SharedSearchState off ExpressionContext (#23331)
# Issue
[SERVER-91262](https://jira.mongodb.org/browse/SERVER-91262)

# Description

Move the SharedSearchState off the ExpressionContext class and into
DocumentSourceInternalSearchIdLookup class (and rename more aptly to
SearchIdLookupMetrics). Then this metric is exposed as a shared pointer
with a getter and setter, to share with the
DocumentSourceInternalSearchMongotRemote class (so these metrics can
eventually be used for batch size tuning). The metrics class is linked
between the two classes during the desugar of a DocumentSourceSearch
class - into the two aforementioned classes).

A shared_ptr should be appropriate here to avoid memory leaks as the
reference counting will destruct the metrics class whenever both
DocumentSource classes go out of scope, in no particular order.

# Testing

Still in progress (PR is not ready for final review). I'm still running
integration tests and thinking of any other unit tests are appropriate.

[Updated
🌲](https://spruce.mongodb.com/version/666cb5348503060007d88160/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)