SERVER-83524 Enable indexed plans for queries referencing system variables (#23235)
This commit reverts the portion of SERVER-74264 which removed the
ability of agg expression constant folding to optimize
ExpressionFieldPaths representing system variables.

This commit also fixes a bug introduced by SERVER-87159, which
initializes system variables only if they are referenced after parsing.
Before this commit, we initialized them after optimization, but in order
for the aforementioned optimization to work, we need the initialize to
occur after parsing and before optimization.

This allows references to system variables to be constant folded and
thus expressions to be pushed down into the query layer, so that we are
able to generate index bounds for them.