SERVER-91254: Compute threshold as double when value is double (#23652)
In windowsFunction in classic we always convert the base value and the
delta to decimal and compute the threshold of the window as a decimal
value. When later we compare the value against the threshold to decide
whether the document is inside the window we convert the value and the
threshold to double and compare two double values. Converting a double
to/from decimal might result in losing precision and when the delta is
much smaller than the value we might end up trying to remove a document
that was never added in the window. With this change we avoid converting
from double to Decimal to double by computing the threshold as double
when the value is a double.

Co-authored-by: Foteini Alvanaki <foteini.alvanaki@mongodb.com>