SERVER-91671: Set up slack channel alerting for block-on-red (#23668)
SERVER-91671

The CLI was already being called by its "monitor_build_status" task in
the "STM Daily Cron" variant.
This work sends a slack message to a new "#redness-pressure-gauge"
channel. Right now, it's sending verbatim reports, but I expect we'll
tweak it as we figure out what the thresholds and computations will be
exactly in the near future.