SERVER-90395: Do not override clone() for a couple of stages that mongostream supports. (#22960)
# Issue

[SERVER-90395](https://jira.mongodb.org/browse/SERVER-90395)

# Description

Do not override DocumentSource::clone() for `DocumentSourceMatch` and
`DocumentSourceSingleDocumentTransformation`. The overrides were anyways
unnecessary as they were doing the same thing the base implementation in
DocumentSource::clone() does. Removing these overrides allows ASP team
to assert that `clone()` is not overridden for the stages ASP supports
and that these stages continue to support
`SerializationOptions.serializeForCloning` option used in the
DocumentSource::clone().