SERVER-91675 Make SbExpr::extractExpr() pass 'state.frameIdGenerator' to SBEExpressionLowering (#23674)
When generating EExpression trees, the SBE stage builder supports a
"hybrid" model. Part of the EExpression tree can be generated by
SBEExpressionLowering, and part of the EExpression tree can be generated
directly by code in the stage builder (without using
SBEExpressionLowering).

I noticed that for parts of the EExpression tree generated by
SBEExpressionLowering we're generating frame IDs using one mechanism,
and for parts of the EExpression tree _not_ generated by
SBEExpressionLowering we're using a different mechanism.

One mechanism allocates frame IDs starting at 1, and the other mechanism
allocates frame IDs starting at 101, so for smaller queries there will
never be overlap between the set of IDs generated by one mechanism and
the set of IDs generated by the other mechanism.

However, for larger queries, these two sets of frame IDs may overlap
with each other. When there is overlap between these two sets, it's
possible to end up with an `ELocalBind` which uses given frame ID that
has a descendant `ELocalBind` which also uses that same frame ID. If
this scenario were to occur, it would cause a tassert failure in
CodeFragment::declareFrame() when compile() was called on the
EExpression tree.

This PR fixes this problem by ensuring we use the same frame ID
generator for the entire EExpression tree (both for the parts of the
tree generated by SBEExpressionLowering and for the parts of the tree
_not_ generated by SBEExpressionLowering).