SERVER-91465 Differentiate collection generation mismatch between acquire and restore (#23690)
This change introduces a differentiation at the level of the
acquisitions API between the collection generation changing at
acquisition time vs at restore after a yield. The behavior of the former
remains the same as what it is today, namely resulting in a retryable
StaleConfig error. The behavior of the latter, with this change, is to
throw a non-retryable error (QueryPlanKilled).

This differentiation allows us to catch early in the execution of a
query (or multi-write) that the collection got dropped and then
recreated, rather than relying on the query plan to get this. It also
ensures that multi-writes will not get retried by the router if the
meaning of the collection changes midway during the execution.