SERVER-91616 COLLSCAN instead of using index with collation for regex. With query knob. (#23627)
In this PR we stick to the original definition of
[SERVER-65260](https://jira.mongodb.org/browse/SERVER-65260). Since
we've identified [several scenarios that we cannot anticipate at
planning time in which it would degrade
performance](https://docs.google.com/document/d/1FMcDB1zTs2DxovKlChDPt8XVD7g98BmIcZIk03ThvK0/edit?usp=sharing)
we're adding a query knob as gatekeeping mechanism.

This PR is a mix of reverting https://github.com/10gen/mongo/pull/17968
and changes.