SERVER-91132: Add client side support for GWProxy failure recovery (#22966)
A couple notes:

- The reason for this change is that the MMS team noted that it's not
really practical for us to use Route53 health checks to determine if a
GWProxy node is healthy (and then return the healthy node only in the
DNS query), so we're instead adding all of the available nodes to a
single A record, and then choosing one on the client side.

- Since I need a random result from the answer that getaddrinfo()
provides, I use a random number generator. This is notoriously slow, but
I keep it out of the "fast path", so we'd use it only at connection
setup. If we're not collectively ok with that, I can find a different
way to do this too, but this is probably the right-est way.

- In practice this approach works well. If a GWProxy server fails, we
retry fairly rapidly, and still converge pretty quickly.

- In the next phase of this, we'll introduce some more deterministic
probes for health checking instead.