SERVER-91462 Change the error thrown if accurate percentile requested with feature flag on (#23569)
Added conditional during validation of percentile (and median) such that if the feature flag for accurate percentiles is on and an accurate method such as "discrete" or "continuous" is input, it will produce a error message
that the accurate methods are not implemented. If the flag is off it will error that only "approximate" can be used for percentile. Changes were made to jsTests to distinguish between feature flag on/off errors.