SERVER-88594 Write a test to assert that internal commands have authchecks  (#21385)
This PR includes the following process:

1. Four setups are used: standalone, sharded cluster mongos and mongod
shard and replication sets
2. checkCommandCoverage: For each setup, listCommands is run to get the
set of commands available for this setup. (This is done with the help of
commands_lib.js; commands_lib.js is modified to accommodate internal
commands only options to the checkCommandCoverage).
3. For each command, it verifies that internal commands can only run by
the authorized users by using following two tests
a) If the authorized user runs the command, it wont return authErrCode
or 13.
b) If an un-authorized user runs the command, it returns authErrCode or
13.
4. a) If both tests are successful, then the command Object in the
internalCommandMap is modified for the test object is updated with a
result of SingleCommandResult.Pass. b) if an unauthorized user can run
the command, the test object is updated with a result of
SingleCommandResult.UnauthorizedAccess. c) otherwise the test is
repeated on the next test setup (if any are remaining).
5. Once all the four setups complete their checks, all the internal
commands objects have some results. The results are published with a
summary of commands that ran successfully and commands that are
suspected of allowing unauthorized access.

Latest:

Two commands are skipped:
_shardsvrCoordinateMultiUpdate is failing because it passes with empty
[] role.
_isSelf currently passes the test, but it will get an exception for not
using ['__system'] role.