SERVER-91116 Remove restriction on over-aligned decorations (#22885)
SERVER-91116
Remove restriction on over-aligned decorations

We don't allow over-aligned decorations today/any decorations with an
alignment requirement > std::max_align_t.
A static assert enforcing this was added when it was noticed that the
original decorations implementation might not correctly support
over-aligned decorations.

This PR adds additional support to the decorable library for
over-aligned decorations, and a test to verify that we do correctly
support over-aligned decorations, and removes the static assert.