SERVER-91060 Use correct nodeType for magic restore sharded cluster passthroughs (#23422)
This PR re-enables the magic restore sharded cluster passthroughs. I've
updated the hook to pass in the correct nodeType for sharded clusters.
This means we're running the "update shard metadata" code path for the
first time, so I needed to make a few small fixes to make this work.

This is a duplicate of an older PR that I closed, the difference being
that I removed the `configsvr` and `shardsvr` parameters from the magic
restore invocations in the passthroughs. I also added guardrails to
prevent users from running magic restore with `shardsvr` or `configsvr`
mode, as this'll prevent magic restore code from running. We will only
do magic restore in replica set mode.