SERVER-89890 Fix assertRetryableCommandWorkedOrFailedWithCodes (#23433)
The helper function `sh.assertRetryableCommandWorkedOrFailedWithCodes`
was written in a way that didn't work for all cases. Fixing this issue
led to finding some places where code was relying in the incorrect
behavior. Finally, some unit tests were also updated.

It's worth noting that this is a somewhat hacky solution. The ideal
solution would be to have a helper library that everyone could rely on
that provided a simple a correct way to determine the nature of a
command error: error code, is it retryable, transient, write error, etc.

---------

Co-authored-by: Silvia Surroca <silvia.surroca@mongodb.com>