SERVER-90864 Clear filtering metadata as part of each node's release of the critical section. (#23236)
# Issue

SERVER-90864

# Description

The proposed changeset adds one extra "functor" parameter to the
`ShardRecoveryService::releaseRecoverableCriticalSection()`, allowing
callers to specify an extra action to be executed before the critical
section document is removed; such an action would on the other hand
skipped in case the critical section specified on the request is already
released.

The extra actions supported are "No Action" and "Clear Filtering
Metadata". The logic of the existing clients of `ShardRecoveryService`
has been adapted to the new API, while preserving the original
semantics.

(Also, secondary nodes will keep reacting to the release of a critical
section by clearing their filtering metadata, based on the existing code
of `shardsvr_op_observer.cpp@onDelete()`).