SERVER-91782 Move kernel version config check under conditional (#23913)
This kernel version check should only be required in certain conditions,
either we are on linux or if on some other platform, the user
specifically requested tcmalloc google.

The kernel version check was running even on windows or other platforms
where the build would not care about kernel version and can produce
misleading error messages.