SERVER-90635 fix ninja bazel libdeps dependencies (#24005)
This change adds support for ninja to see ThinTarget library
dependencies. When ThinTargets were introduced, scons leveraged the
command line subst generator "for_signature" to detect if command lines
changed. The bazel library dependencies are not represented in scons
graph to start reducing scons graph size. In scons the bazel
dependencies are summarized in the for_signature of the command line
generator.

However in ninja the "for_signature" feature is not available. In ninja
we need to express the dependencies in its graph. This is okay for ninja
as its much faster at graph walking and processing.