SERVER-91888 Add checks against repeat fieldnames in bsoncolumnbuilder ref object merge (#24012)
An object merging with a reference object during the construction of an
interleaved sequence in such a fashion that a fieldname was repeated and
as a result checks against an object/array element being delta'ed
against a scalar element were bypassed caused an invariant to fail. The
same checks used for followup elements at the end of the merge are now
added to all locations where we build up the merged result to guard
against this.