SERVER-91964 fix system allocator not working with macos dynamic link build (#24087)
When ThinTargets merged, dynamic link builds for macos broke because of
an exception in a scons subst generator. These exceptions in generators
are hard to debug because scons hides the exception. Upon closer
inspection (adding temporary exception handling to the function), I
found that there was type error of when adding to list together. In some
cases the item ended up being None type, and could not be added with a
list.

For dynamic link builds this is because the shim_allocator has no
dependencies in bazel, not even the global ones that are always there
for other libraries.