SERVER-91913 fix asm fPIC flags (#24091)
CPPFLAGS was the wrong var to use to adjust assembler flags, while it
did adjust assembler flags it also adjusted other types of build
actions, including C and C++ for static and dynamic builds. ASPPFLAGS is
the correct var to use to adjust the .S type assembly files.

This was tested with a required patch build, and a patch build showing
the failing test now green.