SERVER-91465 Make update/delete query plans handle ShardCannotRefreshBecauseLocksHeld by yielding (#23921)
When performing writes to collections undergoing
resharding/moveCollection, shards need the temporary resharding
collection routing table in order to annotate the `destinedRecipient` on
the oplog entry (this is, the shard that would own that document in a
post-resharding world). This is typically done through the
`ShardingWriteRouter` class (e.g.
[here](https://github.com/mongodb/mongo/blob/b13d5b81e3a17a28c0b52a7ec924067439927481/src/mongo/db/op_observer/op_observer_impl.cpp#L930)
for updates).

`ShardingWriteRouter`'s constructor will, when resharding is in certain
stages, [ask the CatalogCache for the temporary resharding collection
routing
table](https://github.com/mongodb/mongo/blob/b13d5b81e3a17a28c0b52a7ec924067439927481/src/mongo/db/s/sharding_write_router.cpp#L71-L75).
When constructed as part of op_observer::onUpdate, locks are being held.
If it happens to be that the CatalogCache does not have the requested
entry currently cached, this would require a refresh, which involves
network requests. This is not acceptable while holding locks, so the
CatalogCache [throws
ShardCannotRefreshBecauseLocksHeld](https://github.com/mongodb/mongo/blob/b13d5b81e3a17a28c0b52a7ec924067439927481/src/mongo/s/catalog_cache.cpp#L419-L420).
Shards handle this exception by tearing down the query plan, dropping
locks, then waiting for the cache to be refreshed and finally
propagating the exception to the router. The router considers
ShardCannotRefreshBecauseLocksHeld as retryable, so it retries. This is
fine for single-writes, but for multi-writes it can cause double-updates
if it had already partially executed.

The proposal in this PR is to have UpdateStage/DeleteStage handle
ShardCannotRefreshBecauseLocksHeld internally by requesting a yield,
instead of failing the operation and bubbling up the exception to the
router. The nss for which a catalog cache refresh is needed is annotated
on a new OpCtx decoration named
`planExecutorShardingCatalogCacheRefreshRequired`. Then, when the query
plan has yielded the locks, if that decoration is set, we'll refresh the
catalog cache before restoring. Upon restore, the query plan can
continue executing normally.