SERVER-90927: Add kafka source offset lag to verbose stats (#23857)
We obtain the kafka source offset lag by using a rdkafka API -
[get_watermark_offsets](https://github.com/confluentinc/librdkafka/blob/bb2843b5ce32c52954fced9b746205c9340bf96f/src-cpp/rdkafkacpp.h#L1642)

This is a local call and returns the cached high offset. The cached
value is updated by librdkafka every time it gets a new message from the
broker.

Added a jstest to see if a non-zero lag is observed. Unlike the change
stream case, we seem to be able to observe a non-zero lag without
needing to introduce an artificial sleep somewhere.