SERVER-92000: Allow stream processor does not exist  error in stop thread of ConcurrentStartStop_StopDuringConnection (#24138)
* Fix for https://jira.mongodb.org/browse/BF-33800
* This unit test occasionally fails in evergreen and locally because the
stop thread fails with `StreamProcessorNotFound`
* This error is expected because, if there is a concurrent stop request,
the executor `connect/start` flow stops, and the processor gets removed
from the map, and the stop request might fail with
`StreamProcessorNotFound`. Logs from the test failures indicate this is
happening: "Executor future returned early during start, likely due to a
connection error or an external stop request"
* The streams Agent already treats a StreamProcessorNotFound error as a
success in the stop flow, so if a user calls start/stop concurrently,
they should still get an `ok: 1` back in this scenario.