SERVER-90511 Support serializing optimized DocumentSourceLookUp (#23826)
This PR is adding support for serializing a DocumentSourceLookUp
instance and re-parsing (createFromBson) it back to create a similar
DocumentSourceLookUp instance. The idea is to serialize an optimized
query plan and support parsing it back to a DocumentSourceLookUp
instance without calling optimize.
 
This change is part of
[SERVER-78464](https://jira.mongodb.org/browse/SERVER-78464) to support
checkpoint for a serialized execution plan for ASP.