SERVER-91685 Make TransactionCoordinator continuation chains keep the object alive (#23698)
This change resolves a race that could occur under very specific
conditions where `TransactionCoordinator::_done` would execute, start
running `_completionPromise.setError`, and before `setError` had
finished, the `TransactionCoordinator` would already be deleted by
another thread.

`TransactionCoordinator` was relying on a contract of "it is illegal to
delete a TransactionCoordinator before waiting on its onCompletion()
future". However, this contract is subtle broken. `setError`, `setFrom`,
and the like can be fully observable from the future side _before_
they've finished executing fully. So, deleting the
`TransactionCoordinator` in response to observing `onCompletion()` isn't
safe as it may delete `_completionPromise` in the middle of it executing
`_completionPromise.setError`.

To fix the race, this change makes the continuation chains that were
previously set up in the constructor keep the `TransactionCoordinator`
alive by capturing `shared_ptr`s obtained from `shared_from_this()`.
This required some minor refactoring:

1. `TransactionCoordinator` now exposes a `start` function that kicks
off asynchronous operations. This is necessary as `start()` needs to use
`shared_from_this`, which generally can't be called from the
constructor.
2. `TransactionCoordinator` now exposes a `shutdown` function that
destroys its schedulers. This is necessary as they must be destroyed
before the parent scheduler, and therefore callers may need precise
control over when it happens. With the coordinator keeping itself alive,
they no longer can rely on it being destroyed around the time of
`onCompletion()` firing (though, I think this was racy to begin with, we
just consistently won the race).
3. I've removed the warnings that it is illegal to delete a
`TransactionCoordinator` before waiting on `onCompletion`. It's no
longer illegal, it's now impossible. The `TransactionCoordinator` will
keep itself alive until it's safe to delete it.

The changes in the tests are mechanical.