SERVER-89145: Always update Kafka consumer group offsets during stop (#23861)
* Currently we update Kafka consumer group offsets after the
corresponding checkpoint is flushed, and we always write a final
checkpoint during stop. But prior to this PR, we were not always
updating the kafka consumer group offsets for the final "stop"
checkpoint. This is important to fix because a customer wants to stop
their processor and rely on the Kafka consumer group offsets to
accurately reflect the latest offset processed.
* The solution is to wait for the final "stop" checkpoint to be flushed
before the executor finishes shutting down.