SERVER-86051 move recorded mongo version to archive_dist_test_debug (#24197)
Currently record_mongodb_server_version.sh runs `mongod --version` in
compile_dist_test. The problem with this is that segfaults can happen
during mongod startup. If this segfault happens in compile_dist_test
then it happens before the binaries/debugsymbols are uploaded so we get
a core dump with no binaries to debug it with.

It should be move to the end of archive_dist_test_debug after
debugsymbols are uploaded.