SERVER-90876 Make movePrimaryFailIfNeedToCloneMovableCollections only allow movePrimary to move unsharded collections that are not movable by moveCollection (#23701)
We want to ensure that only collections that cannot be tracked are
allowed to be moved by movePrimary in the config transition suites. We
are specifically checking for `isNamespaceAlwaysUntracked` and making an
exception for `system.resharding` collections so that they can be moved
by movePrimary, as they can be generated by the fuzzer.