SERVER-90319 Refactor semaphore ticket holder unittests (#21892)
This patch adds some new unittests for the SemaphoreTicketHolder, and
restructures their fixture so the tests are deterministic/don't rely on
"soon" behavior. Summary of all the new tests and what they cover:

- QueuedWaiterGetsTicketWhenMadeAvailable -> tests that a blocked waiter
is woken up when an in-use ticket is returned to the pool

- TicketsAreDistributedFifo -> tests that when in-use tickets are
returned to the pool, they are distributed to waiters in the order they
began queueing

- QueuedWaiterRemainsNegativeAfterReleaseToPoolWithNegativeAvailable -
checks blocked waiter remains queued when a ticket is released, but the
available count is still negative

- CanResizePool - basic test that we can resize the pool up and down
- ResizeDownTicketsStillAvailable - Tests that when we reduce the
capacity of the ticketholder but _tickets is still positive, tickets can
still be checked out

- ResizeDownSoNoTicketsAvailable - Tests that we can reduce the capacity
of the ticketholder below the number of currently checked out tickets,
meaning that negative tickets are available. Tests that tickets can't be
acquired after this downsizing.
- ResizeUpMakesTicketsAvailableToWaiters - tests that a blocked waiter
is woken up when the pool is re-sized such that tickets are now
available

- ResizeUpDistributedTicketsInFifoOrder -tests that when multiple
waiters are waiting, and the pool is resized to make more tickets
availalbe, those new tickets are distributed in FIFO order to waiters.
Additionally, tests the number of waiters woken up corresponds to the
newly availalbe number of tickets (and others remain asleep).