SERVER-92037 Fix data race in AuthorizationSessionTestFixture using an atomic (#24198)
There is a data race in the `AuthorizationManagerImpl::_authEnabled`
flag when
`UnauthorizedSessionIsCoauthorizedWithAnybodyWhenAuthIsDisabled` and
`AuthorizedSessionIsCoauthorizedNobodyWhenAuthIsDisabled` call
`authzManager->setAuthEnabled(false)` outside of initialization time. At
the same time, the `JournalFlusher` thread is reading its value through
`isAuthEnabled()`.

This is an unintended use of `_authEnabled`, which currently allows
unsynchronized accesses because it's only expected to be set once at
initialization time. To allow these unit tests (and future tests) to
safely reset its value, `_authEnabled` is made an atomic.