SERVER-90213 Implement the replication waiter list as a table of lists (#23877)
Implement the replication waiter list as a table of lists to improve
efficiency under load

This replaces the WaiterList used for _replicationWaiterList with a
WriteConcernWaiterList, which tracks each write concern separately so it
is not necessary to go through the entire list on every wake-up. Thus in
cases where we have many waiters whose conditions are not yet fulfilled,
those unfulfilled waiters will not slow down wakeups of those which are.

Since we're not going through all the waiters on wakeup, the delayed
remove mechanism cannot be used, so instead the remove methods now
require opTime to be given so the to-be-removed waiter can be found
quickly.