SERVER-86736: Tune mongot cursor batchSize on lookup id success rate (#24204)
[JIRA Ticket](https://jira.mongodb.org/browse/SERVER-86736)

**Description**
The `SearchIdLookupMetrics` are passed into the
`MongotTaskExecutorCursorGetMoreStrategy` so that as the
`DocumentSourceInternalSearchIdLookup` updates its search metrics,
subsequent batch sizes to mongot can be tuned accordingly. Specifically,
for queries where the limit is extractable, and the query is not stored
source, the success rate of finding documents returned by mongot is used
to influence the guess of how many more doucments should be requested
from mongot.

**Testing**
Unit tests have been added to mongot_cursor_test.cpp &
mongot_cursor_helpers_test.cpp to test that the batch size updates
properly as the `SearchIdLookupMetrics` are updated, for the right type
of queries.

---------

Co-authored-by: Will Buerger <59492746+wbuerger46@users.noreply.github.com>