SERVER-92065 Limit the number of applyOps operations in the oplog application buffer. (#24395)
The new vectored inserts project groups inserts into applyOps oplog
entries. Each oplog entry can hold up to 500 applyOps operations.
Limiting the buffer only by number of oplog entries is not sufficient to
bound the lag between lastApplied and lastWritten, especially when there
are a lot of indexes on the collection. So this patch puts a 10k limit
to the number of applyOps operations in the oplog application buffer
instead.